Support custom suspendable annotation
All references to `@Suspendable` now also check for a custom annotation
along side it.

Custom annotation is set via a static function
`Classes.setCustomAnnotation`.

This _should_ be passed as an argument in the `JavaAgent` which would
then set the static variable. Allowing it to be accessed by the
instrumenting code.